Ingress_v2/requirements.txt
shodan>=1.31.0
python-dotenv>=1.0.0
folium>=0.16.0
rich>=13.0.0
